Refactor: src폴더별로 index.js를 만들어 import 구문 단순화
