+++ src/Validator/CarValidator.js
@@ -1,6 +1,5 @@
-import { CAR_CONSTANTS, SYMBOL } from '../Constants/Constants';
-import { ERROR_MESSAGES } from '../Constants/Messages';
-import AppError from '../utils/Error';
+import { CAR_CONSTANTS, ERROR_MESSAGES, SYMBOL } from '../Constants';
+import { AppError } from '../utils';
 
 const { NAME_LENGTH_RANGE } = CAR_CONSTANTS;
 const { INVALID_NAME_LENGTH, DUPLICATED_NAME, INCLUDES_SPACE_NAME } = ERROR_MESSAGES;
@@ -8,8 +7,7 @@ const { INVALID_NAME_LENGTH, DUPLICATED_NAME, INCLUDES_SPACE_NAME } = ERROR_MESS
 const CarValidator = (() => {
   const checkNameLength = (carNames) => {
     const isOverLength = carNames.some(
-      (carName) => carName.length > NAME_LENGTH_RANGE.max
-        || carName.length < NAME_LENGTH_RANGE.min,
+      (carName) => carName.length > NAME_LENGTH_RANGE.max || carName.length < NAME_LENGTH_RANGE.min,
     );
     if (isOverLength) {
       throw new AppError(INVALID_NAME_LENGTH);
