+++ src/Validator/index.js
@@ -1,0 +1,3 @@
+export { default as CarValidator } from './CarValidator';
+export { default as CommonValidator } from './CommonValidator';
+export { default as TryNumValidator } from './TryNumValidator';
