+++ src/View/InputView.js
@@ -1,6 +1,5 @@
 import readLine from 'readline';
-import { VIEW_MESSAGES } from '../Constants/Messages';
-import { SYMBOL } from '../Constants/Constants';
+import { SYMBOL, VIEW_MESSAGES } from '../Constants';
 
 const { PROMPT_CAR_NAMES, PROMPT_TRY } = VIEW_MESSAGES;
 const { DIVIDE_SYMBOL } = SYMBOL;
