+++ src/utils/index.js
@@ -1,0 +1,2 @@
+export { default as AppError } from './Error';
+export { default as pickRandomInt } from './pickRandomInt';
